app/python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.33.0
mysql-connector-python==9.2.0
